Added Image Filter and Noise cancellation tasks

+++ celery_test/add_run_tasks.py
@@ -14,7 +14,8 @@ f = os.listdir(upload_folder)[0]
 fh = open(file_path+"/celery_test/Uploads/"+f)
 num_lines = sum(1 for line in fh)
 res_out = np.zeros((num_lines,),dtype=bool)
-os.unlink(output_folder+"/"+os.listdir(output_folder)[0])
+if os.listdir(output_folder) :
+    os.unlink(output_folder+"/"+os.listdir(output_folder)[0])
 start = time.clock()
 start1 = time.time()
 start2 = timer()
@@ -27,21 +28,12 @@ for f in os.listdir(upload_folder):
     for line in fh:
         a,b = line.split(" ")
         result = longtime_add.delay(int(a),int(b))
-        # print(result.task_id)
-        # # at this time, our task is not finished, so it will return False
-        # print('Task finished? ', result.ready())
-        # print('Task result: ', result.result)   
-        # # sleep 10 seconds to ensure the task has been finished
-        # time.sleep(10)
-        # # now the task should be finished and ready method will return True
-        # print('Task finished? ', result.ready())
-        # print('Task result: ', result.result)
+        
         ressult.append(result)
         outt = [t.get() for t in ressult]
 
-        
-        f_out.write(str(result.result)+"\n")
         if result.ready()==True:
+            f_out.write(str(result.result)+"\n")
             res_out[c] = True
         c = c+1
